Refactored DriverFactory
